handler: avoid overwriting the connection data pointer
For compatibility with future nginx versions, we shouldn't overwrite
the connection object's data pointer.

Modify the code to use a module-specific context pointer instead of
overwriting c->data. This allows c->data to remain a pointer to the
ngx_http_request_t object and makes the request pointer in
ngx_http_fd_pass_ctx_t redundant.

Fixes invalid access of c->data as an ngx_http_request_t object in
ngx_http_set_lingering_close(). Note that lingering close should not
be used for connections that have been handed off to a different
process; this will be fixed in an upcoming commit.

Signed-off-by: Kenneth Klette Jonassen <kenneth@bridgetech.tv>